Move logic for handling diverging functions and panics calls out of Function
These call belong in special_function.rs. These are cases that do not concern the active blocks and should be treated closer to a real function call. These makes more sense semantically. An unwind terminator is not the same as a call to `panic!`. Now we can make a difference. Modified the tests accordingly.