Make the mapping between local variables and mutexes part of the specific function being translated
The previous implementation used a global mapping between locals and mutex references which meant that that it did not support mutexes in different functions.

Since `rustc_middle::mir::Local` is an index valid only for the current function, we need to keep a mapping of locals to mutexes for every function.