Implemented multithreading: Spawning and joining threads
Use a similar idea to MutexManager. We keep track of the variables that contain join handles as we do with the lock guards for the mutexes.

The thread functions/closures are translated after the main thread finishes. We store the information we need to translate the thread as a `ThreadSpan`.

The translation loop for the thread function is similar to the translation loop for the main thread.

See `test_minimal_thread_spawn.rs` for an example of a resulting Petri net.