Refactor: Create separate enum variants in FunctionCall for mutex functions.
The logic of translating a `Mutex::new()` is quite different from translating `Mutex::lock()`. We end up passing a lot of arguments down to the MutexManager when we could simply separate the paths at the level of `prepare_function_call`.

Now this is implemented. There are separate methods for translating each function call.