Implemented handling of the `Unreachable` terminator.
This generator is useless for the translation but it shows up in MIR representations of for loops working with iterators.
In order to support for loops and other constructs, we need to provide a reasonable solution.

The solution in this case is connecting the basic block to the PROGRAM END place to prevent triggering a false deadlock or a false panic.