Refactor: Use FunctionCall as an enum signaling the function type. Move the handlers to a separate submodule.
There is no need to make the enum hold copies of all the information needed for the function call if the information is going to be used right away in the next method call.

We need an enum to abstract finding the right handler for the function. Then we can simply call the corresponding handler for the enum type. An intermediate object `FunctionCall` is pointless.

This change simplifies the Translator considerably as most methods now live in function_call.rs