Important refactor: Use `Place` for the mapping in Memory instead of `Local`.
The local variabels (_1, _2, _3) may have different subfields. By storing the Place instead of the `Local` we abstract ourselves away from this complexity. It will allow us to map variables like _6.0 and _6.1 to two different mutexes in a seamless  way. This is necessary to implement moving variables to new threads as the closure stores the moved variables as fields of the local variable containing the closure.

This is an important architecture change because a lot of the code relies on memory.rs to keep track of mutexes, lock guards and join handles. On the other hand, using `Place` forces us to define the explicit lifetime <'tcx> in many places to ensure that the place lives as long as the compilation process itself.

The nice thing is that not having to work with `Local` simplifies many functions. Now the usual way of getting the type of a variable is through `Place::ty`.