Fix: Separate the unwind flow from the successful flow when calling a foreign function.
This is a long outstanding bug: The transition representing a function call cannot model a disjunction in the execution flow. We need a second transition for the cleanup place.

Refactor the labels for the functions in mutex, arc and thread too. It is not necessary to use hardcoded strings in the methods when we can move that to naming submodules.