Refactor: Implement error handling for arc creation using an interface to the `netcrab` library
First of all there was a "label duality". There were labels for the places and transitions but also different labels for the panic messages. This was confusing and cumbersome to work with.
Now the labels used for the Petri net (the ones defined in naming) are also used for the error messages. This removes the need for a lot of static strings dispersed around the code.

Using a single entry point to the `add_arc_place_transition` and `add_arc_transition_place` allows us to modify the imports to rely solely on the new Petri Net interface to obtain what we need.

error_handling.rs is no longer needed. Remove it and move the last &str to translator.rs