Refactor: Relax type checks to include more cases when tracking variables
Instead of checking for an exact match, we can test for the substring to be present. This makes the code support tuples in arbitrary order with mutexes, condvars, join handles, etc.
Matching every possible case with an if case if not feasible.
We also cannot keep track of every possible container where these variables may be stored. But we can check for the presence of the type and treat the variable as a mutex/condvar/join handle/mutex guard if that is the case.