Implement a new way of passing mutexes from one thread to the other using VarDebugInfo
This way works with the new nightly compiler.
We can rely on the fact that the debug info will always be there to let us "connect" the mutex created in one thread to the local variable of the new thread.

The debug info is also available when compiling in release mode, which is great news.