Refactor: Generalize the logic in ArcManager to handle clone and dereferencing other sync primitives.
First of all we should check if the contents of the Arc is a condvar, mutex or a combination.
The logic is analogous to handle_assignment, so we can reuse this function.

On the other hand, a variable may be a mutex AND a condvar at the same time, i..e. a tuple containing both. So the if-else cases are actually NOT exclusive. We need to check every possible case.

Lastly, the expected substring of the Mutex and MutexGuard overlap causing a variable to be marked as a mutex and then as a mutex guard. This is incorrect. By adding the "<" to the expected substring we can solve this behavior in an elegant way.