Refactor: Simplify the logic of ArcManager
ArcManager does not need to handle the translation of side effects if the logic is always the same. We can define a method in sync.rs and use it in function_call.rs directly.

This makes the ArcManager a mere counter. Its logic can be moved to a more general structure for this purpose.