Refactor: Convert ArcManager to FunctionCounter
Now indexed functions (functions with labels that are unique for every call) are handled in a more general way. There is no need for an ArcManager as the logic of FunctionCounter can be reused.

Simplify the naming submodules. Move the contents of arc.rs to function.rs.

Rename the Counter to HashMapCounter to make the semantic clearer.