Fix major regression in the MIR related to the introduction of the UnwindAction enum.
This is the relevant PR that introduced the changes to rustc
https://github.com/rust-lang/rust/pull/102906/

Now there a new enum called `UnwindAction`, which may contain in some cases a basic block for the cleanup that we need to handle.

This change requires some renaming to keep some consistency when using the words "unwind" and "cleanup"

All tests are passing again.

Updated the README