Refactor integration tests using a macro to significantly reduce boilerplate
Now every file is just a simple call to a declarative macro that generates 3 test functions (one for every format) from the given paths.

We considered implementing a more general macro that scans all the files in `examples/` but then we would lose the capabilities of running/skipping specific tests passing parameters to `cargo test`.