Refactor of the integration tests and the CLI
To support executing multiple tests concurrently, we need to create the files in a different place that the usual `./net.dot`.
In order to do this we need to support setting a different output folder via the CLI.
For the tests, we would like to output the files right next to the expected results, but this mean that we cannot use the name `net` anymore.
So then support for a different filename was added.

Now the integration tests are defined in files that match the folder names in `examples/programs`.
Since submodules in the `tests` folder are not supported (see https://doc.rust-lang.org/book/ch11-03-test-organization.html), this is the best we can do to somehow "match" the folder structure. This does require more boilerplate but less files overall.