Number all foreign function calls by default
This is crucial to support programs that call the same function many times in different places (not in a loop for example).

The tests are all green. Nonetheless this needs a refactor since the logic is very similar to the logic used for the existing functions that deal with synchronization primitives. Those were already indexed.