Refactor of the logic for foreign function calls: Use indexing for ALL calls.
This unifies the handling, simplying considerably the code.

First, there is no need to have separate labelling/naming functions for every sync primitive. We can always use the same labels for every function as long as we make sure that the function name is sanitized correctly.

Second, the cases in `function_call_handler.rs` become exactly equal, so that we can join the match arms.

Third, `function_counter.rs` was not a good idea. We can directly operate with the counter from the Translator itself. There is no need for a separate component. By reducing complex cases to the simple ones, we can reuse logic and make the Translator more straightforward.

All the tests are passing. The changes are only labelling changes.