Improvement: Check ALL foreign function calls for a sync variable
The idea is to perform the check in `link_return_value_if_sync_variable` for all foreign function calls.

This saves us having to maintain a list of "supported functions" function from the std where we actually track if a mutex/lock guard/condvar/join handle is being passed around. It also simplifies the number of cases in FunctionCall that we need to handle.

The new solution is more robust since it always performs the check, making it "future-proof" in a certain way.
In the process, we needed to handle extracting the arguments in a different way that ignores constants and skips the check if the function has zero arguments.