Improvement: Support all combinations of target and unwind action for function calls.
A specific case that was not supported was the function `std::process::exit`, which in some cases generates a cleanup block but no return block.

Besides, the new `match` considers all possibilities and handles the `Unreacheable` case where it was simply ignored before.

The new `match` statement cannot be moved to a separate function because of the divergent function case.