Improvement: handle drop and assert of the form `[return: bbX, unwind terminate]`
This is important since a lot of example programs exhibit this structure since the `UnwindAction` enum was introduced to rustc in https://github.com/rust-lang/rust/pull/102906

In a previous commit, the change in rustc was incorporated but without handling all the cases.
This improvement now handles all the variants that
enum `UnwindAction` has instead of ignoring some of them.
https://github.com/hlisdero/granite2/commit/8cf95cd54b29c210801cae2941abcbbb85051b92