Optimizations regarding the querying and use of `function_name`
In `is_foreign_function` check first for functions from the standard library or core library to shortcircuit the OR quickly.
Receive the `function_name` instead of querying it again.

In `MIRFunction`, receive the `function_name` instead of querying it again.

In `start_function_call`, receive the `function_name` instead of querying it again.

Handle querying the name only in two places: When calling a function and when pushing a function to the stack.