Fix bug: Connect the mutex to the cleanup transition of the `lock` function
This is crucial since otherwise the deadlock will not be detected!
The program could continue down the cleanup path, evading the detection by the model checker.

This is an important design decision. The alternative would be to NOT model the cleanup path, i.e. ignore it. This would mess too much with the MIR representation of the code. If the MIR says that there is a cleanup, we definitely have to model it. But to detect the deadlock, we add an additional artificial edge from the mutex to the cleanup transition. This is therefore a compromise solution for the purposes of this tool.