Architectural change: Do not model MIR statements anymore.
Modelling MIR statements does not add value to deadlock or lost signals detection.
As seen in the [CFG documentation](https://rustc-dev-guide.rust-lang.org/appendix/background.html#what-is-a-control-flow-graph), the statements cannot change the control flow. That can only happen at the terminators.

In all the programs we have tested so far, the statements do not perform any action that may justify their addition to the Petri net. On the contrary, the nets that include the statements are considerably bigger and much more difficult to read.

In order to facilitate the use and adoption of this tool, it is crucial to optimize the Petri net for the purpose of the tool. The decision is to eliminate all the code related to the modelling of the statements and fix the tests accordingly to match the new output.

The alternative was to disable the statements with a compile flag but that would just complicate testing and since there is no use case for modelling the statements anyway, we discarded this option.

This commit can be used as a reference for the future of how to implement a Petri net representation of the MIR statements.