Architectural change: Do not model the `unwind terminate` assert and drop.
tl,dr: It is not relevant to the deadlock analysis and just makes the Petri nets bigger.

They basically represent what to do when a drop or an assert fails, this is an extremely hideous case that does not occur in practice. There is no need to "duplicate" the cleanup paths for the sake of completeness when it makes the translation more difficult to read.