Minor fixes in the call to the model checker
For some reason, the witness path is printed to stdout while the rest of the output is printed to stderr. Nonetheless the output to stdout is consistently empty and I could not find a place in the original LoLA code that explains why the output does not end up there.

I suppose that LoLA is using some non-standard way of printing to stdout that is not being caught by Rust when executing the command.