Assume that `notify_one` never unwinds
There is an issue with how rustc compiles this call to MIR. In some cases, the `std::sync::Condvar::notify_one` function contains a cleanup target.  This target is not called in practice but creates trouble for lost signal detection. The reason is that any call may fail, which is equivalent to saying that the `notify_one`  was never present in the program, leading to a false lost signal.
In conclusion: Ignore the cleanup place, do not model it. Assume `notify_one` never unwinds.