Refactor: Implement MutexRef and MutexGuardRef as shared pointers
This removes the need for a significant amount of reference validity checks during runtime. It replaces costly function calls with less costly pointer dereferences.

It simplifies the `MutexManager` considerably and allows access to the underlying `Mutex` without requiring the `MutexManager`, which makes the API simpler.

See this question for a comparison of the old solution vs the new solution:
https://stackoverflow.com/questions/40875152/reference-to-element-in-vector