Refactor `Memory`: Introduce `MemoryMap`
The logic in `Memory` was repeated for every sync variable possible.
By using a new struct with type parameter, we can extract this logic and generalize it.
After that, we discovered that memory could directly expose the methods in `MemoryMap`, making the API clearer.

Now `Memory` is just a container of 4 `MemoryMap`