Refactor `Condvar`
Rename places and transitions in the model to make their purpose clearer.

Inline the function `translate_call_wait`. It is no longer necessary to separate in cases, making it short enough for an inline.

Lock and unlock the mutex in the method of `Condvar`. It is part of the logic of the sync variable itself.