WIP: First implementation of an extended model of condition variables
The problem with the previous implementation is that nothing stops the model checker from taking the path straight to the call to `wait()` disregarding the value of the flag contained in the mutex.

This situation could happen in ANY case that a condition variable is used.

The first attempt at a solution was working with `wait_while` which incorporates the `wait()` itself and the while loop in a single function call.
This saves us having to track which flag goes with which condition variable and which while loops are relevant for a given condition variable (analyzing while loops is a path we are definitely not going to take..)

Sadly, this does not make the issue any better. The model checker can still fire the transitions it pleases.

We focused then on another issue, namely the modelling of the flag itself. If we can take into account its value, we can prevent the `wait()` from occurring. One case of this is when the variable has already been set, then no `wait()` should be called. Instead, the `wait()` should be skipped. Since the condition variable logic cannot be "permanently" activated (filled with tokens so that it acts as a NOP), we took the approach of trying to disable the usual wait path and open another alternative path called `WAIT_SKIP`.

Unfortunately, this comes with its own set of difficulties. On the one hand, the current implementation tracks the `deref_mut` calls on a mutex guard and connects later the `wait_enable` place to the function transitions, effectively disabling the condition variable. Then, it sets a token on a new `condition` place that is supposed to enable the skip path (that should be disabled when the condition variable is NOT set.
But this last arcs should be set on a global level after translating ALL threads. This is not easy and the Translator currently does not have any kind of postprocessing after translating every MIR function.

The next step was to construct a simpler example to see if the approach of the condition place even works. The program `set_condition_before_wait` has the statements in the exact order that triggers the "correct" translation with the new model.
The result? It does not work yet. There are still cases for which the new model fails.

However the tests pass with minor modifications.