Assume that dereferencing or calling `unwrap` on a variable containing a mutex or a mutex guard never fail
This causes trouble for lost signal and deadlock detection as seen previously. The comments details which assumptions are being made. The resulting nets are a bit smaller, which is always something useful for the analysis.

Fix the tests and add a LoLA test for `wait_while` and `wait`.  Now they both produce the right result thanks to the global postprocessing task of linking the condvar to the mutex.