Architecture change: Add support for aggregates in the `Memory`
This involves changing the hypothesis for the `Memory`. The translation of the programs is not affected. Nonetheless, this change will allow to support more complex programs.

Before, an aggregate value was simply modelled as superposition of multiple values.
For example: `_1` may contain one mutex and one condvar at the same time.
The translator just links (associates) the place to the two different sync variables. That's why we were keeping separate mapping for each kind of sync variable.

This works well as long as there are not any instances of multiple values of the same type in the same aggregate. For instance, trying to pass TWO condvars or TWO mutexes did not work before because the translator just confounds them and overwrites the previous mapping. It cannot tell which is which.

Now it is possible to write programs that pass multiple sync variables of the same type.