Add support for the deadlocks caused by `join`
Now, ignoring the cleanup place of the call to `join`, we may cause a deadlock in the Petri net model.

This showed us that `two_threads_sharing_mutex` actually had a deadlock because it cannot join the thread if the mutex lock is not dropped first.

It also showed that `producer_consumer` has a deadlock because it tries to join threads that do not return.

Additionally there was a typo in the test for `dining_philosophers`. The program DOES have a deadlock and the translator detects it.

All tests passing.