feat: update to `nightly-2023-11-01` and fix regressions
There were a couple of new config fields in `rustc_interface::Config`. The default config suffices for our purposes.

Replace `.get(0)` with `first()` as suggested by clippy.
Rename `thread_function_def_id` in `Thread` as suggested by clippy.

`GeneratorDrop` was removed and a new `CoroutineDrop` was added. No idea if they are used for the same. I do not need them in the implementation currently.