feat: new value implementation
The idea is to track the aggregates no matter how nested they are.

Use a new type of Value `Other` to fill in when creating aggregates that
have multiple types in them.

Check correctly that the return value and the first argument have the same sync var type.
This was a bug in the previous solution. It was assuming that functions that receive a sync var in the first argument always return the same sync var as the return value and therefore a linking between the two should occur.
This was WRONG! Example: Thread spawn receives a closure with a mutex but the return value is a join handle.

Notice that we should trip references when comparing types

We should accept both cases when working with sync variables:
- Unboxing a sync variable (return value is shorter than original type)
- Boxing a sync variable (return value is longer than original type)

Improved type checks in a way that calls fewer types the tcx object for the optimizer MIR. This is more efficient.

Remove thread::spawn from the list of supported functions.
There is no handle for it