feat: search for sync vars in aggregates recursively and use new heuristic for linking return value with first argument
The new condition for assuming that a function is working with a sync variable is the following:

- The return value and the first argument should be of the same type or an aggregate of the same type
- The first argument should contain a linked value

This handles well the case of `thread:spawn` for instance. The first argument is a closure which in most cases contains a sync variable but the return value is a join handle.

As for the aggregates, we need to search for the corresponding type inside in a recursive manner.

This commit passes all the existing tests.