feat: new memory implementation supporting aggregate nesting
The new memory model supports nesting of aggregates, i.e. structs, vecs, and other similar language constructs.

It does not map places. Instead it maps locals. This is closer to how the memory works in MIR functions.
For simplicity, `Memory` implements index-based operations like a `Vec`.

Add a debug function that allows to print the whole memory of a function easily and spot issues.

As a nice side effect:

- The lifetime 'tcx is no longer needed in many places
- Moving the sync variables to the thread does not rely on the debug info. This makes the solution more robust. The new solution is straightforward.

All existing tests pass with this new memory implementation