feat!: implement linking return value of MIR function to its destination
This required a significant refactor of how the return values are
handled. To streamline the setup, I decided to make every function
return a `Option<Value>`. The MIR visitor handle then links if a `Value`
is present.

This partly deprecates the old heuristic that checks if the first
argument is of the same type. Instead of that, we know which kind of
function we are translating, so we can return the value so that the MIR
visitor handler links it to the destination at the end.
**In the case of unknown functions of the std library, the heuristic is
still applied although in a simplified way**.

All the test are passing, including a better version of the Dining
Philosophers problem.